search tv shows or movies on selecting the corresponding tab
